midaircommit

+++ src/main/java/no/nav/bidrag/beregn/bidragsevne/beregning/Bidragsevneberegning.java
@@ -1,5 +1,6 @@
 package no.nav.bidrag.beregn.bidragsevne.beregning;
 
+import java.math.BigDecimal;
 import no.nav.bidrag.beregn.bidragsevne.bo.GrunnlagBeregningPeriodisert;
 import no.nav.bidrag.beregn.bidragsevne.bo.ResultatBeregning;
 
@@ -7,10 +8,10 @@ public interface Bidragsevneberegning {
     ResultatBeregning beregn(
         GrunnlagBeregningPeriodisert grunnlagBeregningPeriodisert);
 
-    Double beregnMinstefradrag(
+    BigDecimal beregnMinstefradrag(
         GrunnlagBeregningPeriodisert grunnlagBeregningPeriodisert);
 
-    Double beregnSkattetrinnBelop(
+    BigDecimal beregnSkattetrinnBelop(
         GrunnlagBeregningPeriodisert grunnlagBeregningPeriodisert);
 
     static Bidragsevneberegning getInstance(){
