+++ src/main/kotlin/no/nav/bidrag/beregn/bidragsevne/bo/periode.kt
@@ -5,11 +5,12 @@ import no.nav.bidrag.beregn.felles.bo.PeriodisertGrunnlag
 import no.nav.bidrag.beregn.felles.enums.BostatusKode
 import no.nav.bidrag.beregn.felles.enums.InntektType
 import no.nav.bidrag.beregn.felles.enums.SaerfradragKode
+import java.math.BigDecimal
 
 data class InntektPeriode(
     val inntektDatoFraTil: Periode,
     val inntektType: InntektType,
-    val inntektBelop: Double) : PeriodisertGrunnlag {
+    val inntektBelop: BigDecimal) : PeriodisertGrunnlag {
   constructor(inntektPeriode: InntektPeriode) : this(inntektPeriode.inntektDatoFraTil.justerDatoer(), inntektPeriode.inntektType,
       inntektPeriode.inntektBelop)
   override fun getDatoFraTil(): Periode {
